requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==5.3.0
etelemetry==0.3.1
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.12.2
fitz==0.0.1.dev2
fsspec==2023.1.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==4.13.0
isodate==0.6.1
joblib==1.3.2
looseversion==1.3.0
lxml==5.3.2
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.6
nltk==3.8.1
numpy==1.21.6
packaging==24.0
pandas==1.3.5
pathlib==1.0.1
Pillow==9.5.0
prov==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pyparsing==3.1.4
pypdf==5.0.0
PyPDF2==1.26.0
python-dateutil==2.9.0.post0
python-multipart==0.0.8
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.1
rdflib==6.3.2
regex==2024.4.16
requests==2.31.0
safetensors==0.4.5
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.27.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
traits==6.3.2
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.15
uvicorn==0.22.0
zipp==3.15.0